modify vmem type int to ULL

+++ src/processinfo.h
@@ -50,7 +50,7 @@ struct ProcInfo {
     int     ppid;       // parent process id
     string  comm;       // command
     double  cpu;        // cpu 사용률
-    int     vmem;       // 메모리 점유율
+    ULL     vmem;       // 메모리 점유율
     char    state;      // state
     string  user;       // 프로세스 사용 유저명
     string  start;      // 프로세스 시작 시간
